# Optimize proto translation for merger