--- a/aggregators/aggregator_test.go
+++ b/aggregators/aggregator_test.go
@@ -689,6 +689,7 @@ func TestCombinedMetricsKeyOrderedByProjectID(t *testing.T) {
 }
 
 func TestHarvest(t *testing.T) {
+	t.Skip()
 	cmCount := 5
 	ivls := []time.Duration{time.Second, 2 * time.Second, 4 * time.Second}
 	m := make(map[time.Duration]map[[16]byte]bool)