--- a/aggregators/merger.go
+++ b/aggregators/merger.go
@@ -38,9 +38,6 @@ func (m *combinedMetricsMerger) MergeOlder(value []byte) error {
 }
 
 func (m *combinedMetricsMerger) Finish(includesBase bool) ([]byte, io.Closer, error) {
-	// TODO: Investigate test failures. Releasing this resource is causing test failures
-	// Are we holding a reference to released resource?
-	// defer m.metrics.ReturnToVTPool()
 	data, err := m.metrics.MarshalBinary()
 	return data, nil, err
 }